Added functions 'loadArticles()' and 'saveArticles()' to load and save 'Article' objects as dictionaries in a json file.
This uses the methods 'ent_to_dict()' and 'dict_to_ent()' which convert 'feed.entries' dictionary objects containing time.struct_time objects into purely string based dictionaries which can be saved.  Currently, before saving, the title of articles are checked against those already in the save file --- to see if they are new.  Something more robust needs to happen.  Also, the save file is a list of dictionaries.  It would probably be better to just have a dictionary per line --- so that new entries can easily be appended (if we know they are new, based on e.g. datetime information).